+++ cli/azd/test/functional/cli_test.go
@@ -8,6 +8,7 @@ import (
 	"context"
 	"crypto/rand"
 	"encoding/hex"
+	"encoding/json"
 	"fmt"
 	"io/fs"
 	"io/ioutil"
@@ -23,7 +24,10 @@ import (
 
 	"github.com/azure/azure-dev/cli/azd/cmd"
 	"github.com/azure/azure-dev/cli/azd/internal"
+	"github.com/azure/azure-dev/cli/azd/pkg/azureutil"
+	"github.com/azure/azure-dev/cli/azd/pkg/commands"
 	"github.com/azure/azure-dev/cli/azd/pkg/environment"
+	"github.com/azure/azure-dev/cli/azd/pkg/osutil"
 	"github.com/azure/azure-dev/cli/azd/pkg/project"
 	"github.com/azure/azure-dev/cli/azd/test/azdcli"
 	"github.com/joho/godotenv"
@@ -128,6 +132,101 @@ func Test_CLI_Init_CanUseTemplate(t *testing.T) {
 	require.FileExists(t, filepath.Join(dir, "README.md"))
 }
 
+// Test when we have multiple resource group matches. More than one rg has azd-env-name set
+func Test_CLI_ResourceGroupNameWithMultipleMatches(t *testing.T) {
+	envName := randomEnvName()
+	Internal_Test_CLI_ResourceGroupsName(t, envName, fmt.Sprintf("rg-%s", envName), true, true, true)
+}
+
+// Test when we can't find any resource group matches
+func Test_CLI_ResourceGroupNameWithoutMatch(t *testing.T) {
+	envName := randomEnvName()
+	Internal_Test_CLI_ResourceGroupsName(t, envName, fmt.Sprintf("rg-%s", envName), true, false, false)
+}
+
+// Test when resource group uses rg- prefix
+func Test_CLI_ResourceGroupNameWithPrefix(t *testing.T) {
+	envName := randomEnvName()
+	Internal_Test_CLI_ResourceGroupsName(t, envName, fmt.Sprintf("rg-%s", envName), true, true, false)
+}
+
+// Test when resource group uses -rg suffix
+func Test_CLI_ResourceGroupNameWithSuffix(t *testing.T) {
+	envName := randomEnvName()
+	Internal_Test_CLI_ResourceGroupsName(t, envName, fmt.Sprintf("%s-rg", envName), true, true, false)
+}
+
+// Test when we don't have any resource groups with azd-env-name tag
+func Test_CLI_ResourceGroupNameWithoutEnvNameTag(t *testing.T) {
+	envName := randomEnvName()
+	Internal_Test_CLI_ResourceGroupsName(t, envName, fmt.Sprintf("rg-%s", envName), false, true, false)
+}
+
+func Internal_Test_CLI_ResourceGroupsName(t *testing.T, envName string, rgName string, includeEnvNameTag bool, createResources bool, createMultipleResourceGroups bool) {
+	ctx, cancel := newTestContext(t)
+	defer cancel()
+
+	dir := t.TempDir()
+	t.Logf("DIR: %s", dir)
+
+	//envName := randomEnvName()
+	t.Logf("AZURE_ENV_NAME: %s", envName)
+
+	cli := azdcli.NewCLI(t)
+	cli.WorkingDirectory = dir
+	cli.Env = append(os.Environ(), "AZURE_LOCATION=eastus2")
+
+	// Store the original environment to be used later
+	originalEnvironment := cli.Env
+
+	err := copySample(dir, "resourcegroups")
+	require.NoError(t, err, "failed expanding sample")
+
+	cli.Env = append(originalEnvironment, fmt.Sprintf("TEST_RESOURCE_GROUP_NAME=%s", rgName))
+
+	if includeEnvNameTag {
+		cli.Env = append(cli.Env, "TEST_INCLUDE_ENV_NAME_TAG=true")
+	}
+
+	if createMultipleResourceGroups {
+		cli.Env = append(cli.Env, "TEST_CREATE_MULTIPLE_RESOURCE_GROUPS=true")
+	}
+
+	_, err = cli.RunCommandWithStdIn(ctx, stdinForTests(envName), "init")
+	require.NoError(t, err)
+
+	if createResources {
+		_, err = cli.RunCommand(ctx, "infra", "create")
+		require.NoError(t, err)
+	}
+
+	envFilePath := filepath.Join(dir, environment.EnvironmentDirectoryName, envName, ".env")
+	env, err := environment.FromFile(envFilePath)
+	require.NoError(t, err)
+
+	// Verify that resource group is found or not found correctly
+	foundRg, err := azureutil.FindResourceGroupForEnvironment(ctx, &env)
+
+	if createResources {
+		if createMultipleResourceGroups {
+			// We have multiple resource groups, so we expect an error
+			require.Error(t, err)
+		} else {
+			// We found a single resource group, so we do not expect an error
+			require.NoError(t, err)
+			require.Equal(t, foundRg, rgName)
+		}
+
+		// Using `down` here to test the down alias to infra delete
+		_, err = cli.RunCommand(ctx, "down", "--force", "--purge")
+		require.NoError(t, err)
+	} else {
+		// We didn't create the resources, so we expect an error
+		require.Error(t, err)
+	}
+
+}
+
 func Test_CLI_InfraCreateAndDelete(t *testing.T) {
 	ctx, cancel := newTestContext(t)
 	defer cancel()
@@ -151,12 +250,20 @@ func Test_CLI_InfraCreateAndDelete(t *testing.T) {
 	_, err = cli.RunCommand(ctx, "infra", "create")
 	require.NoError(t, err)
 
-	file, err := ioutil.ReadFile(filepath.Join(dir, environment.EnvironmentDirectoryName, envName, ".env"))
+	envFilePath := filepath.Join(dir, environment.EnvironmentDirectoryName, envName, ".env")
+	env, err := environment.FromFile(envFilePath)
 	require.NoError(t, err)
 
 	// AZURE_STORAGE_ACCOUNT_NAME is an output of the template, make sure it was added to the .env file.
-	// (the name of the resource is ${replace(envName, '-', '')}sa).
-	require.Regexp(t, regexp.MustCompile(fmt.Sprintf(`AZURE_STORAGE_ACCOUNT_NAME="%sstore"`, strings.ReplaceAll(envName, "-", ""))), string(file))
+	// the name should start with 'st'
+	accountName, ok := env.Values["AZURE_STORAGE_ACCOUNT_NAME"]
+	require.True(t, ok)
+	require.Regexp(t, `st\S*`, accountName)
+
+	// Verify that resource groups are created with tag
+	rgs, err := azureutil.GetResourceGroupsForEnvironment(ctx, &env)
+	require.NoError(t, err)
+	require.NotNil(t, rgs)
 
 	// Using `down` here to test the down alias to infra delete
 	_, err = cli.RunCommand(ctx, "down", "--force", "--purge")
@@ -164,8 +271,6 @@ func Test_CLI_InfraCreateAndDelete(t *testing.T) {
 }
 
 func Test_CLI_InfraCreateAndDeleteWebApp(t *testing.T) {
-	t.Skip("flakiness in app service preventing infrastructure creation")
-
 	ctx, cancel := newTestContext(t)
 	defer cancel()
 
@@ -199,14 +304,30 @@ func Test_CLI_InfraCreateAndDeleteWebApp(t *testing.T) {
 	url, has := env["WEBSITE_URL"]
 	require.True(t, has, "WEBSITE_URL should be in environment after infra create")
 
-	res, err := http.Get(url)
-	require.NoError(t, err)
+	// Add a retry here because appService deployment can take time
+	err = retry.Do(ctx, retry.WithMaxRetries(10, retry.NewConstant(5*time.Second)), func(ctx context.Context) error {
+		t.Logf("Attempting to Get URL: %s", url)
 
-	var buf bytes.Buffer
-	_, err = buf.ReadFrom(res.Body)
-	require.NoError(t, err)
+		res, err := http.Get(url)
+		if err != nil {
+			return retry.RetryableError(err)
+		}
 
-	assert.Equal(t, "Hello, `azd`.", buf.String())
+		var buf bytes.Buffer
+		_, err = buf.ReadFrom(res.Body)
+		require.NoError(t, err)
+
+		testString := "Hello, `azd`."
+		bodyString := buf.String()
+
+		if bodyString != testString {
+			return retry.RetryableError(fmt.Errorf("expected %s but got %s for request to %s", testString, bodyString, url))
+		} else {
+			assert.Equal(t, testString, bodyString)
+			return nil
+		}
+	})
+	require.NoError(t, err)
 
 	// Ensure `env refresh` works by removing an output parameter from the .env file and ensure that `env refresh`
 	// brings it back.
@@ -307,8 +428,14 @@ func Test_CLI_InfraCreateAndDeleteFuncApp(t *testing.T) {
 	err = cmd.Execute([]string{"deploy", "--cwd", dir})
 	require.NoError(t, err)
 
-	fnName := envName + "func"
-	url := fmt.Sprintf("https://%s.azurewebsites.net/api/httptrigger", fnName)
+	out, err := cli.RunCommand(ctx, "env", "get-values", "-o", "json", "--cwd", dir)
+	require.NoError(t, err)
+
+	var envValues map[string]interface{}
+	err = json.Unmarshal([]byte(out), &envValues)
+	require.NoError(t, err)
+
+	url := fmt.Sprintf("%s/api/httptrigger", envValues["AZURE_FUNCTION_URI"])
 
 	t.Logf("Issuing GET request to function\n")
 
@@ -430,14 +557,14 @@ func copySample(targetRoot string, sampleName string) error {
 		targetPath := filepath.Join(targetRoot, name[len(sampleRoot):])
 
 		if info.IsDir() {
-			return os.MkdirAll(targetPath, 0755)
+			return os.MkdirAll(targetPath, osutil.PermissionDirectory)
 		}
 
 		contents, err := ioutil.ReadFile(name)
 		if err != nil {
 			return fmt.Errorf("reading sample file: %w", err)
 		}
-		return ioutil.WriteFile(targetPath, contents, 0644)
+		return ioutil.WriteFile(targetPath, contents, osutil.PermissionFile)
 	})
 }
 
@@ -466,9 +593,12 @@ func getTestEnvPath(dir string, envName string) string {
 // the provided `testing.T` has a deadline applied, the returned context
 // respects the deadline.
 func newTestContext(t *testing.T) (context.Context, context.CancelFunc) {
+	ctx := context.Background()
+	ctx = context.WithValue(ctx, environment.OptionsContextKey, &commands.GlobalCommandOptions{})
+
 	if deadline, ok := t.Deadline(); ok {
-		return context.WithDeadline(context.Background(), deadline)
+		return context.WithDeadline(ctx, deadline)
 	}
 
-	return context.WithCancel(context.Background())
+	return context.WithCancel(ctx)
 }