+++ ext/vscode/src/telemetry/activeSurveys.ts
@@ -1,0 +1,9 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+import { SurveyOne } from "./surveyOne";
+import { Survey } from "./surveyScheduler";
+
+export const activeSurveys: Survey[] = [
+    SurveyOne
+];